pyproject.toml
[project]
name = "dtp-auth"
version = "0.1.0"
description = "FastAPI authentication backend for the DT platform"
readme = "README.md"
authors = [
    { name = "Yin-Chi Chan", email = "ycc39@cam.ac.uk" },
    { name = "Anandarup Mukherjee", email = "am2910@cam.ac.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.119.0",
    "jwt-pydantic>=0.0.7",
    "psycopg>=3.2.10",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sortedcontainers>=2.4.0",
    "sqlmodel>=0.0.27",
]

[tool.uv.build-backend]
module-name = "dtp.auth"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest>=8.4.2",
]
